Expand `suggest_dataset_name` to potentially suggest multiple
I noticed that for the prompt `populations` it was suggesting `hospital-codes`, this was because there were multiple 'options' that were equally 'close'. This tweak allows suggesting any amount when many are equally close.

I don't think `hospital-codes` is at all a good suggestion for `populations` but to fix that we'd have to change how it works out the string distances (e.g. change the `method`).