Use RETRY instead of GET
This should mean that any failed requests will be retried up to 3 times on a failure. The function will wait a random amount of time, progressively longer on each retry, so as not to overwhelm the server.

It should clear up the intermittent errors we get.

Possibly we want to use `quiet = TRUE` - at the moment it will print a message when it has to retry informing how long it will wait before trying again.