Address the R CMD CHECK NOTE in newer R versions
This check is not in our current R version (4.1.2) but is included in the newer versions. As a result, it only shows up in the CI jobs. For example: https://github.com/Public-Health-Scotland/phsopendata/actions/runs/13679978038/job/38250546098#step:6:94

Although the job isn't failing since it's just a 'NOTE,' I believe CRAN expects there to be no NOTES or, at the very least, an explanation for them. Fortunately, in this case, resolving the issue is quite straightforward.